zephyr: implement console read/snapshot host commands
Implement an optional console output buffer, and wire it up to the
associated host commands.

We don't use the EC's uart buffering module for this, as it implements
a UART buffer which also is used for transmit buffering (we don't do
that, and instead immediately send to printk).

We don't use Zephyr's ring buffer structure for this as we need to
track multiple heads due to the semantics of snapshots.

BUG=b:178033156
BRANCH=none
TEST="ectool console" on volteer

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: Ib1d39e32e8bad5b23b98034c33f56534a8c89fcb