flash_ec: look for npcx_monitor.bin relative to EC image
Given a local EC build, npcx_monitor.bin can be found at a well known
location relative to ec.bin (chip/npcx/spiflashfw). So, add that to the
search path. flash_ec was failing to find npcx_monitor.bin when only
using the --image flag for a local build:

  util/flash_ec --image build/${BOARD}/ec.bin

as it was looking for npcx_monitor.bin in the same directory as ec.bin.
The layout in /build/${BOARD}/firmware/... has ec.bin and
npcx_monitor.bin in the same directory, but a local build in the EC tree
uses build/${BOARD}/chip/npcx/spiflashfw/npcx_monitor.bin.

Also, adding the ${BOARD} environment variable as the default board type
when --board is not specified. This matches the behavior of the
makefiles. In this case, we can invoke flash_ec without arguments.

BRANCH=none
BUG=b:132350402
TEST=used flash_ec to update the EC on brya

Change-Id: I669f52cc4fe480ae23914d9fcbe67aadd18ed6c3
Signed-off-by: Caveh Jalali <caveh@chromium.org>