Avoid passing char values to ctype functions
With Zephyr's newlib we get a warning when trying to do this, since it
includes the following note in:

  /opt/zephyr-sdk/arm-zephyr-eabi/arm-zephyr-eabi/include/ctype.h

  "These macros are intentionally written in a manner that will trigger
  a gcc -Wall warning if the user mistakenly passes a 'char' instead
  of an int containing an 'unsigned char'."

Presumably this is so characters above ASCII 127 are handled correctly,
even if these are seldom used.

Update the few affected call sites to ensure the value is cast to an
unsigned char so that it will not fall afoul of the newlib warning.

Note that the ECOS version of the ctype functions does not make use of
an array, so does not suffer from failure if negative values are
passed. Still, it is harmless to fix it.

BUG=b:180023514
BRANCH=none
TEST=build zephyr volteer with CONFIG_NEWLIB_LIBC and see no warnings

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: Ieae2fab8c20b75baa46d01dd8cdb393c6bb5c2c2