BB retimer: Put the BB retimer in LPM on exit disconnect
In the event we have the retimer firmware update mechanism
ready by M90, we may get into a situation where the first
boot after M90 will not have the retimer FW update applied,
and so for that first boot, mode switching will not work

Hence, as a work around from the first boot -
Reset the retimer on disconnect but putting it in low power mode
and then bring back at connect on entering USB mode.

BUG=b:178025846
BRANCH=None
TEST=Tested with older retimer fw version - Able to enter into
alternate mode.

Signed-off-by: Ayushee Shah <ayushee.shah@intel.com>
Change-Id: I6f1b8d2e398589cc8f654bfa28ee75560a742711