magolor and friends: Adjust OCPC PID constants
Thus far, it seems that boards that use the same OCPC PID constants
behave similarly.  Therefore, this commit updates the constants that
were most recently used on madoo to the other OCPC boards that use the
same charger IC such that the behaviour meets the criteria at
go/ocpc-testing.

- Converges around the target current.
- Average steady state error is less than 4% of target.
- Current is reduced to target level in less than 5s when load is
  suddenly released.

BUG=None
BRANCH=dedede
TEST=Build and flash magolor. Verify that battery can charge to 100%.
Verify that it meets or exceeds the constraints above.
TEST=Verify that DUT can come out of battery cutoff from sub board.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: I206854c097c307b941a64547f9b74c8259a7d499