TCPMv2: DPM: Add support for sending SVDM Attention message
This CL adds a new DPM function to generate a request to the PE to
send a SVDM Attention message via the DPM. This change deprecates
pd_send_hpd from TCPMv2, which existed for the same
purpose. pd_send_hpd is now part of the hpd->DP Attention converter.

This path replaces the need to use pd_send_vdm which can't be reliably
used with its current design.

BUG=b:175660576,b:180465870
BRANCH=None
TEST=Verified on quiche that can enter ALT-DP mode as a UFP-D and that
display is extended properly via display port or hdmi connector.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I9f8817cb8d40f32b878abdc09db605097a32548b