zephyr: zmake: Simplify finding the modules
There are two types of modules: those that are part of Zephyr and the
EC shim, which is in the EC source tree. It is painful to try to use
the same API for these.

In addition, the 'modules' argument to locate_modules() is never used.
It isn't really possible to use it since we cannot specify its
dictionary value easily on the command line.

All the modules are in the same directory at present, both in the
Chromium OS tree and in Zephyr. We can require that this continues.
Add a --modules-base flag which provides that directory.

Use the EC directory as the shim directory, since it always is. Use the
current directory as the EC directory, since that seems to be how we
use it. We could always add another argument to zmake to build source
from a different directory if needed.

Simplify locate_modules() while we are here. It is easiest to deal with
the two cases separately. Also the ec_dir @property disappears to
nothing so we might as well remove it.

BRANCH=none
BUG=b:178731498
TEST=pytest zephyr/zmake/
zmake configure -B /tmp/z/vol zephyr/projects/volteer/ -t zephyr &&
    zmake build /tmp/z/vol
    (which fails, fixed in the next CL)
in the chroot:
FEATURES=test sudo -E emerge -q zephyr-build-tools
zmake testall

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I1f62e9d9c8e548c1b5e5238e03ad741c29bbd9fc