zephyr: zmake: Always used the supplied build directory
If --build-dir is provided it seems better to always use it. The
current logic in resolve_build_dir() assumes that the build will
progress far enough that we recognise the build directory. If it
doesn't, then the flag is ignored and we build in a default location.

It isn't clear to me why that is better than always using the
directory. Drop this unnecessary logic.

BRANCH=none
BUG=b:178731498
TEST=pytest zephyr/zmake/
zmake configure -B /tmp/z/vol zephyr/projects/volteer/ -t zephyr &&
    zmake build /tmp/z/vol
in the chroot:
FEATURES=test sudo -E emerge -q zephyr-build-tools
zmake testall

Signed-off-by: Simon Glass <sjg@chromium.org>
Change-Id: I9e46d91e79cc946016cf8d33ed6ea1f3862c977e