zephyr: zmake: Delay finding the checkout
If we are given the paths to the modules and zephyr itself, we don't
need a checkout directory. This works in a similar way to the CL that
was reverted earlier. It calculates the directory only if needed.

I haven't attempted to put the code in a common function, since there
might be a nicer way of doing this anyway.

BRANCH=none
BUG=b:178731498
TEST=pytest zephyr/zmake/
zmake configure -B /tmp/z/vol zephyr/projects/volteer/ -t zephyr &&
zmake build /tmp/z/vol
in the chroot:
FEATURES=test sudo -E emerge -q zephyr-build-tools
zmake testall

Change-Id: I0ad1ed1217ec0a22cc78ef5ffbcb413b05af5c00
Signed-off-by: Simon Glass <sjg@chromium.org>