usb_mux: Send missing latest mux state in G3 wake
Unable to enumerate a USB4/TBT device when it's disconnected and
hotplugged at G3.

When the AP comes up after G3 and since the disconnect flag is set,
EC sends disconnect info to the AP probe functions. But the DUT is
still in low power mode, the AP doesn't receive the current mux state.

So, this CL makes sure that the AP receives the latest mux state by
sending it after the SCI mask is enabled.

BUG=b:180377259
BRANCH=None
TEST=USB4 device disconneted and connected in G3, it enumerates after
     the DUT is up

Change-Id: I25926de70f6a0a65aa350f9d862331fe299049d2
Signed-off-by: Madhusudanarao Amara <madhusudanarao.amara@intel.corp-partner.google.com>
Signed-off-by: Ayushee Shah <ayushee.shah@intel.com>