TCPMv2: Don't reset current limit during an FRS
During an FRS, we need to retain the current limit that was given in the
sink_capabilities until an explicit contract is in place. Normally on
entry to Attached.SRC, we will reset the current limit to the default.
Create an exception to not do this during an FRS.

BUG=b:180434465
TEST=With a 3A device loaded with at least 1.5A but less than 3A, make
sure we don't overcurrent the port duruing a fast role swap
BRANCH=None

Signed-off-by: Eric Herrmann <eherrmann@chromium.org>
Change-Id: Ifc332ed0ae25a4f96acd5c0b30cf2941bacdd155