common: add board_is_usb_pd_port_present function
This adds a new overridable function to check if a USB PD port exists on
a system. This is preferred to using board_get_usb_pd_port_count or
CONFIG_USB_PD_PORT_MAX_COUNT because it handles sparse port numbering.
The "missing" port no longer needs to be the highest numbered port.

BRANCH=none
BUG=b:173575131

Change-Id: I3f74eddd99c6901b42ce05bab6f2bdd545127d1a
Signed-off-by: Caveh Jalali <caveh@chromium.org>