zephyr: zmake: rename "raw" packer to "npcx" packer
The "raw" packer used to be a binary packer which simply forwarded
zephyr.bin to the output directory with no modifications.

A history of CLs has repurposed this packer to be specific to binman
RO/RW format, with the RO expected to generate a Nuvoton loader at the
beginning of the image.  So it's hardly "raw" anymore.

Rename it to "npcx" so we can bring back the original "raw" packer,
needed by chameleon.  (Note this is done in a follow-up CL).

BUG=b:180545676
BRANCH=none
TEST=zmake testall

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I73e13426d7c87db1ccf3514dd321640ebc625068