usb_mux: Don't set force disconnect mux flag in G3
Unable to enumerate a USB4/TBT device when it is hotplugged at G3.

When the AP comes up after G3 and since the disconnect flag is set,
EC sends disconnect info to the AP probe functions. But the DUT is
still in low power mode, the AP doesn't receive the current mux state.

So, this CL makes sure not sending the disconnect mode instead send
latest mode. In G3 wake disconnect mode configuration is not required.

BUG=b:180377259
BRANCH=None
TEST=Connected USB4 device connect in G3 and checked enumeration

Signed-off-by: Madhusudanarao Amara <madhusudanarao.amara@intel.corp-partner.google.com>
Change-Id: I466a9bdbf568c95fe8cd655e25b558d81c904007