zephyr: support zephyr v2.5 on all projects
Zephyr v2.5 will build for all projects now, mark it as supported for
all projects and it will be enabled by default.

We can delete v2.4 from this list once v2.5 gets a bit more battle
testing.

If you need to build for v2.4 in the interim, you should pass the
--zephyr-base flag to zmake, like so:
$ zmake --zephyr-base ~/trunk/src/third_party/zephyr/main/v2.4 ...

BUG=b:180409973
BRANCH=none
TEST=zmake testall
TEST=flash on delbin, test UART console functioning

Signed-off-by: Jack Rosenthal <jrosenth@chromium.org>
Change-Id: I9cb7902455b2c6f231bea5c0b08b0e18bca0ddd8