TCPMv2: Reset discovery events on AP reset
When the AP resets, reset the SOP and SOP' discovery events so it knows
there is completed discovery information to consume.  Additionally,
remove the requirement that events only be set from the port's PD task.
Since this is just an atomic or of events and the setting of a host
event, it should be safe to do from other tasks.

BRANCH=None
BUG=b:179505381
TEST=on galtic, verify events are reset after reboot with:
- Apple 3-in-1 (SOP identity, no SOP')
- TBT dock (both SOP and SOP' identity)
- Active cable tied to another chromebook (SOP' identity, no SOP)

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: Ib2c49854fd4269c65e771565210fe88222c6e848