driver/it83xx, it8xxx2: turn on PD BMC PHY for TCPMv1
If explicit contract is set in bbram when EC boot up, then TCPMv1
set soft reset as first state instead of unattached.SNK, so we need
to enable BMC PHY for tx module.
And same as TCPMv2, we turn on/off BMC PHY and set low power mask
when hook_dis/connect is called.

BUG=none
BRANCH=none
TEST=on icarus, the explicit contract is set in bbram when EC boot up,
     and we can tx soft reset (won't happen tx timeout).

Signed-off-by: Ruibin Chang <ruibin.chang@ite.com.tw>
Change-Id: I3c8d3db240d2290ced17b90aeb0420ac9f20598d