TCPMv2: Don't check VBUS when entering BIST mode
Temporarily remove check for VBUS = vSafe5V when entering BIST mode. It
does this by measuring VBUS, but this facility is inadequate to the
purpose on Volteer. Since this even more of an edge case than actually
running BIST tests, remove the check for now, pending a more general
solution.

BUG=b:180957710,b:173023378
TEST=make buildall; pass TDA.2.1.2.2 BMC PHY RX INT REJ on voxel
BRANCH=firmware-volteer-13672.B-main

Signed-off-by: Abe Levkoy <alevkoy@chromium.org>
Change-Id: I172c1bcc10731c2e405344cdc6ad2b5a1ed80ced