TCPMv2: Fix check for svdm response w/ Attention
Attention messages do not require a response message. A previous CL
had a check for this message in a helper function. That method was not
correct because the the vdm message buffer in the PE object does not
get cleared and so this buffer should only be checked in a state where
the vdm buffer contents are known to be current.

BUG=b:175660576,b:173027965
BRANCH=None
TEST=Verified on quiche that can enter ALT-DP mode as a UFP-D and that
display is extended properly via display port or hdmi connector.

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: Id9908fd4bd3db574fb6f769ab0d2e1db0be5aecd