RAA489000: Rearrange inits to sink Vbus faster
In order to more quickly detect and sink Vbus, move up the priority on
enabling the Vbus ADC and checking our Vbus status to enable sinking.
Additionally, gate this enablement on the battery not being capable of
supporting the board to prevent premature Vbus sinking in normal
use-cases.

BRANCH=None
BUG=b:178981107,b:178728138
TEST=on storo and sasukette, confirm board can boot with no battery and
TCPC works as expected
On galtic, confirm sinking is enabled on a no-battery boot but not
automatically enabled with a charged battery attached

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I26eec09c2fa50d2425c5d6b0ef7b8847bd570da7