TCPMv2: Add PD interface to asynchronously request TC Error Recovery
If a condition is reached outside the PD task which requires USB Type-C
Error Recovery, it can now use pd_set_error_recovery() to request that
the PD task execute this.  Note that this will return before the error
recovery has completed, which is in contrast to the suspend port
behavior.

BRANCH=None
BUG=b:180453483
TEST=make -j buildall

Signed-off-by: Diana Z <dzigterman@chromium.org>
Change-Id: I487e0da3ef65c201ff406500f0c95244b460afcb