panicinfo: fix panicinfo double report
Panicinfo is double report due to the panic flags is reset
whenever a hard-reset triggered. This makes AP thinks there
is a new panic happened, and reports it to the crashmon.

We've seen a lot of crash reports due to the false alarms.
The solution is recording the panic flags across hard-reset
so we won't keep alarming the old panicinfo.

BUG=b:181108236
TEST=1)test that on juniper with new EC RO/RW, and AP reboot
       won't cause the EC alarming a new panicinfo.
     2)test that on juniper with soft-sync RW, and old RO, and
       AP reboot won't cause the EC alarm a new panicinfo
BRANCH=kukui

Change-Id: I6e0c33a1fcf210d9aa03ebd1cd79b84a7744c386
Signed-off-by: Eric Yilun Lin <yllin@chromium.org>