zephyr: Require new CONFIG options to be in Kconfig too
Add a script to help deal with new ad-hoc CONFIG options being added
to ECOS.

This works by maintaining a list of allowed ad-hoc CONFIG options and
refusing to allow a build to succeed if it adds new ones.

This is easier to implement than a pre-submit and more useful, since
we can check the actually CONFIG file produced by a board. For
example, adding a new CONFIG to board/volteer/board.h will cause the
volteer build to break, but not kukui. That seems important to avoid
frustration.

Add the base list of allowed ad-hoc CONFIGs, which should never grow.

BUG=b:181253613
BRANCH=none
TEST=manually test that adding a new '#define CONFIG' causes an error
to be generated, and adding a new Kconfig option with the same name
causes the error to go away.

Also check that the script suggests updating the list when a Kconfig is
added for an existing CONFIG option.

Finally, manually check that ALLOW_CONFIG=1 does as intended.

(that's quite a few manual tests; possibly it is worth automating
them)

Change-Id: Id11b8d859fd07c2db73324cbb9e425182710ab3d
Signed-off-by: Simon Glass <sjg@chromium.org>