TCPMv1/v2: Add config option to disable TCPC VCONN
Currently we always source VCONN from the TCPC, and also the PPC if
present. However the SYV682 can't handle 5V on its host-side CC pins, so
we shouldn't source VCONN in that case.

For those TCPCs which will trigger OVP if VCONN isn't enabled, this will
not happen with the SYV682, since it isolates VCONN from the TCPC CC
signals.

BUG=b:180973460
TEST=On delbin, make sure the host-side CC pins are <3.3V when sourcing
VCONN
BRANCH=None

Signed-off-by: Eric Herrmann <eherrmann@chromium.org>
Change-Id: I8929a44fc23f93c44559229f1b0d024fd7fbc7db