TCPMv2: PE: Add mechanism to request source power role
This CL adds a trigger when evaluating source cap messages to request
a power role swap (via DPM request) that can be customized on a per
board and per port if necessary. The default setting is 'false' so
this change only affects boards that override
pd_check_port_requests_source().

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I9c5bdb77405a6075de9c5b4269213f810a20cc6c