dooly: Fix TCS3400 configuration
On dooly, TCS3400 has an interrupt line to the EC, so we don't need to
simulate it.
When we simulate and an interrupt line is present, 2 interrupts will be
generated (one for a valid sample via GPIO, the other at preset
timeout). When the timeout interrupt pops, there are no sample reday, so
the EC will wait for a one for a long time and prevent other sensors
data to be gatherd.

BUG=b:177860358
BRANCH=puff
TEST=With an android AP pulling sensor data, check messages in
/var/log/cros_ec.log like
0218/071812.170000 [45076.224405 tcs3400_post_events RGBC not valid (0x4)]
0218/071812.174000 [45076.228719 Screen Accel Missed 23 data collections at 2126184833 - rate: 16000]
Are gone.

Signed-off-by: Gwendal Grignou <gwendal@chromium.org>
Change-Id: I213ab02000a030572ed15e34b44a4761df615e2c