zephyr: use interrupt based RX buffering
Poll based RX was technically incorrect since Zephyr was setting up
the UART to FIFO mode (instead of byte mode). This change replaces
polling the RX UART with interrupt based callbacks. When calling
uart_shell_stop(), the RX will be forked and buffered. The buffer can
then be read via the uart_getc() function. It can also be cleared via
the uart_clear_input() function.

It is important to note that using signals was attempted but ended up
inflating the built image too much such that it didn't fit on flash.
It may be revisited later.

BRANCH=none
BUG=b:181352041
TEST=Build volteer, run, see expected 0xec07 from the GSC.

Signed-off-by: Yuval Peress <peress@chromium.org>
Change-Id: I5d2b61e914b56f678a259b373969522da87e8df3