TCPMv2: Don't cflush in CC_OPEN with battery
In the TC state machine, on entry to CC_OPEN we disable vbus/vconn,
print a message, flush the console, then disable CC.

But this takes too long to meet the tProtErrHardReset, which causes some
compliance failures. The main culprit is the console flush. The
justification for the console flush is that we may brown out, so this
isn't necessary when we have a battery.

Only do the console flush on entry to CC_OPEN when we don't have a
battery.

BUG=b:181053528
TEST=Compliance test TD.PD.FRSISNK3.E3 on Voxel
BRANCH=None

Signed-off-by: Eric Herrmann <eherrmann@chromium.org>
Change-Id: Ia056f3c3010737e18c02a7a80659da201cfe5f92