TCPMv2: Change PD_T_PS_SOURCE_OFF value
PD_T_PS_SOURCE_OFF is 750 to 920 ms, currently it is set to the maximum
value. But, we need to detect the state and take the exit action within
this maximum time, which leaves no room for overhead if this is set to
the maximum value.

Instead, set to the midpoint value.

BUG=b:181055620
TEST=TD.PD.FRSISNK3.E5 on Voxel
BRANCH=None

Signed-off-by: Eric Herrmann <eherrmann@chromium.org>
Change-Id: I4b2b75332774ca0fb8e0913053739da6d58d428a