zephyr: Fix version command
The current config_chip used by Zephyr defaulted to setting the RO
header size and offset to 0 if the header was turned off. This isn't
correct as the CONFIG_PLATFORM_EC_RO_HEADER just tells us whether or
not to generate the header, whereas the _SIZE and _OFFSET values
should still be there even in the RW image.

Now running `version` from both the RO and RW image shows:

21-03-02 21:28:21.455 Chip:    chromeos emu
21-03-02 21:28:21.460 Board:   1
21-03-02 21:28:21.486 RO:      _v2.0.7727+3f8b76fa7
21-03-02 21:28:21.496 RW:      _v2.0.7727+3f8b76fa7
21-03-02 21:28:21.497 Build:   _v2.0.7727+3f8b76fa7
21-03-02 21:28:21.500          2021-03-02 21:01:09

BRANCH=none
BUG=b:181599007
TEST=built/flashed volteer and ran `version` from both RO and RW
     images

Signed-off-by: Yuval Peress <peress@chromium.org>
Change-Id: I435f224fc2b3b3a5ca54d3335ea8cd50e9ab5a2e