bb_retimer: Do not talk to burnside bridge when powered off
This adds extra checks to avoid talking to the burnside bridge when we
know it is powered down. Also, do not return EC_ERROR_NOT_POWERED from
the init function when the AP is off as this indicates to the caller
that the chip should be reinitialized to wake up. Since we don't
directly control power to the chip in this case, we should not use this
error code.

BRANCH=none
BUG=b:181743576
TEST=no more I2C errors talking to burnside bridge

Change-Id: Ib43e6ba6ccef022fff6d054203f8490db6c32f42
Signed-off-by: Caveh Jalali <caveh@chromium.org>