zephyr: fix AP power cycling on sysjump
When running the sysjump, the AP was power cycling. This was caused
by the jdata address no longer being correct since the swap to Zephyr
2.5. Previously, the sram0 device tree entry defined a register size
of 0x10000, which needed to have 0x800 bytes removed as part of the
boot ram size. This can be seen in
https://github.com/zephyrproject-rtos/zephyr/blob/v2.4-branch/dts/arm/nuvoton/npcx7m6fb.dtsi#L35

In Zephyr 2.5, the sram0 size was corrected to 62k to account for the
boot ram size. As can be seen in
https://github.com/zephyrproject-rtos/zephyr/blob/master/dts/arm/nuvoton/npcx7m6fb.dtsi#L21

This is also true for npcx7m6fc and npcx7m7fc.

BRANCH=none
BUG=b:181599787
TEST=build and falsh volteer, see correct value returned for
     system_jumped_to_this_image()

Signed-off-by: Yuval Peress <peress@chromium.org>
Change-Id: Icf063ec220d17ebe6a012506e4626702bdf9e149