TCPMv1/v2: Move SOP' enabling to tcpm_set_vconn
Currently SOP' enabling is done as part of the TCPCI driver when vconn
is set - however if we aren't using VCONN from the TCPC, we need to
enable SOP' separately. So, instead of enabling it in the TCPCI driver,
enable it in the general TCPM set VCONN function.

BUG=b:181692098,b:181691263,b:173459141
TEST=Make sure cable discovery works
TEST=make buildall
BRANCH=None

Signed-off-by: Eric Herrmann <eherrmann@chromium.org>
Change-Id: Iecc06760f2b8af588c427b9565c6aa31ee719edf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/ec/+/2733574
Reviewed-by: Abe Levkoy <alevkoy@chromium.org>
(cherry picked from commit 8fffb7f40b72742551a343209a97a7a9add6a827)