ampton: Change input voltage only when conditions changed
Ampton reduces input voltage to 5 volts when battery is full and
chipset is in one of suspend states. When this condition is not true
'reduce_input_voltage_when_full()' always removes the limit. When
battery is charging this function is called every SOC change. As a
result any voltage limit is removed on SOC change which effectively
breaks firmware_PDVbusRequest test.

This patch restricts voltage limit remove to case in which limit was
applied and we decide it is no longer needed.

BUG=b:161775827
BRANCH=none
TEST=Run firmware_PDVbusRequest test on Ampton when battery is not full,
     test should pass.

Signed-off-by: Patryk Duda <pdk@semihalf.com>
Change-Id: I4a7b1db53b89845e63cbb6a1a4927a2d8e4f0a26