util/ec_parse_panicinfo: Parse older structures
For devices on older FW branches, we need to be able to parse
panicinfo even if the size of the structures don't match.

Realign trailer (struct_size and magic) for that purpose.

Also, add WARNING/ERROR tags to messages so that we can
grep for those in a test.

Make a similar change in ectool (use the return value that
indicates the size).

BUG=b:182087232
TEST=make build/bds/util/ec_parse_panicinfo
TEST=build/bds/util/ec_parse_panicinfo < \
      ./upload_file_eccrash-857e3693bfce1e1f.eccrash
BRANCH=none

Change-Id: I97650aa3085e753b0a2256058402e33d11238f52
Signed-off-by: Nicolas Boichat <drinkcat@chromium.org>