TCPMv2: PE/DPM: Don't preemptively set modal flag
Currently we are setting the modal flag as a DFP when the ENTER
command is being sent, without confirmation of the ACK reply from the
UFP. This CL keeps this behavior for TCPMv1, but moves the setting of
the modal flag into the alt mode specific state machine. Currently, we
only support DP and Thunderbolt alt modes.

When the modal flag is set pre-emptively, we will fail the TDA.2.2.3:
BMC-PROT-SEQ-DRSWAP compliance test. This is because we trigger a
hard reset if the modal flag is set and we receive a data role swap
request.

BUG=b:180764800,b:159946351,b:181386294
BRANCH=None
TEST=Verifed that TDA.2.2.3: BMC-PROT-SEQ-DRSWAP passes on quiche

Signed-off-by: Scott Collyer <scollyer@google.com>
Change-Id: I4ae527903859e8ffdb01f355606c5d0f7d177def