charge_state_v2: Guard system_hibernate call with CONFIG_HIBERNATE
This change adds an ifdef guard for CONFIG_HIBERNATE around the call to
system_hibernate. This is to ensure that boards are not forced to
provide the implementation of `system_hiberate` if CONFIG_HIBERNATE is
not selected. It allows early bring up boards to enable charger
functionality without enabling hibernation at the same time.

BUG=b:173575131
BRANCH=None
TEST=make -j buildall

Signed-off-by: Furquan Shaikh <furquan@google.com>
Change-Id: Idcccd3904e9b25f6aa4e06dc90a4845a1a87b2c4