brya: Add workaround for CR50 double reset
This change enables workaround required for cr50 double reset. Copying
commit message from dedede:

Hold the EC on the initial power on until the CR50 resets it; this
preserves the power on flag for the first boot, which is important
when we are deciding if we want to reset the USB-C PD contract.

BUG=b:173575131
BRANCH=None
TEST=make -j buildall

Signed-off-by: Caveh Jalali <caveh@chromium.org>
Signed-off-by: Furquan Shaikh <furquan@google.com>
Change-Id: I2b3411fc540e54ecc185b46f7f9d7e3bb1075847