brya: Add some AP related configs
This change enables the following AP related configs:
1. CONFIG_CMD_AP_RESET_LOG: Enables support for reporting reset logs
to AP.
2. CONFIG_HOSTCMD_AP_RESET: Enables host command that AP can use to
request EC to trigger chipset reset. This was used on boards like
Dedede and Volteer to address a cr50 strap change issue. This issue
might not really apply to brya, but having this host command available
shouldn't hurt.

BUG=b:173575131
BRANCH=None
TEST=make -j buildall

Signed-off-by: Caveh Jalali <caveh@chromium.org>
Signed-off-by: Furquan Shaikh <furquan@google.com>
Change-Id: I44fda98999a93a3bb4a5414adc18c5d734dd6a4d