charger: Add new APIs for measured charge values
Some of the charger ICs can provide measurements for the charge
voltage and the charge current.  This information is needed by the
OCPC module.  Previously, the charge_get_current() and
charge_get_voltage() functions were modified to provide this
information.  However, those functions are intended to provide the set
voltage and current targets for the charger IC.

This commit adds a new set of APIs, charge_get_actual_current() and
charge_get_actual_voltage() which provides the actual charge current
and voltage if the charger IC is able to provide that information.

BUG=b:182018616
BRANCH=dedede
TEST=Build and flash madoo, verify that `charger` EC console command
shows the set current and voltage targets instead of the measured
values.  Check that the `chgstate` command shows the measured values
for use with the OCPC module.
TEST=Verify that charging from the sub board works.
TEST=Verify that resistances are still calculated and seem reasonable.

Signed-off-by: Aseda Aboagye <aaboagye@google.com>
Change-Id: I82565d18908d9ea0f54934787897937488e280e6